Fixed document button style problem and rewritten useSubscription.ts

+++ components/Document.tsx
@@ -2,7 +2,7 @@
 
 import { deleteDocument } from "@/actions/deleteDocument";
 import useSubscription from "@/hooks/useSubscription";
-import { ArrowRightCircle, DownloadCloud, Trash2Icon } from "lucide-react";
+import { ArrowRight, DownloadCloud, Trash2Icon } from "lucide-react";
 import { useRouter } from "next/navigation";
 import { useTransition } from "react";
 import { Document, pdfjs, Thumbnail } from "react-pdf";
@@ -27,7 +27,7 @@ function DocumentViewer({
   const { hasActiveMembership } = useSubscription();
 
   return (
-    <div className="relative flex flex-col w-72 h-80 rounded-xl bg-gray-100 border dark:bg-gray-600 drop-shadow-md justify-evenly p-4 transition-all transform hover:border-indigo-600/40 dark:hover:border-indigo-600/80 cursor-pointer group">
+    <div className="relative flex flex-col w-72 h-80 rounded-xl bg-gray-100 border dark:bg-gray-600 drop-shadow-md justify-evenly p-4 transition-all transform cursor-pointer group">
       <div className="flex flex-col">
         <p className="font-semibold line-clamp-2">{name}</p>
         <p className="text-sm text-gray-500 dark:text-gray-300">{size}</p>
@@ -38,7 +38,7 @@ function DocumentViewer({
         <Document
           loading={null}
           file={downloadUrl}
-          className="h-full w-full overflow-hidden rounded-lg"
+          className="h-full w-full overflow-hidden rounded-lg hover:border-indigo-600/40 dark:hover:border-indigo-600/80"
         >
           <Thumbnail
             className="shadow-lg flex justify-center items-center"
@@ -62,45 +62,50 @@ function DocumentViewer({
         </Button>
 
         <Button
-          disabled={isDeleting || !hasActiveMembership}
+          title={!hasActiveMembership ? "PRO feature" : "Delete"}
+          className={`flex items-center justify-center w-8 h-8 p-0 ${
+            !hasActiveMembership
+              ? "bg-gray-400"
+              : "bg-transparent hover:bg-white"
+          } rounded-full shadow-none border`}
+          disabled={isDeleting}
           onClick={() => {
-            const prompt = window.confirm(
-              "Are you sure you want to delete this document?"
-            );
-            if (prompt) {
-              // delete document
-              startTransition(async () => {
-                // Delete the document from Firestore
-                await deleteDocument(id);
-              });
+            if (!hasActiveMembership) {
+              alert("Upgrade to delete documents");
+            } else {
+              const prompt = window.confirm(
+                "Are you sure you want to delete this document?"
+              );
+              if (prompt) {
+                // delete document
+                startTransition(async () => {
+                  // Delete the document from Firestore
+                  await deleteDocument(id);
+                });
+              }
             }
           }}
-          title="Delete"
-          className="flex items-center justify-center w-8 h-8 p-0 bg-transparent rounded-full shadow-none border "
         >
           {isDeleting ? (
             <span className="loading loading-spinner text-primary"></span>
           ) : (
             <Trash2Icon className="h-4 w-4 text-red-500 z-50" />
           )}
-          {!hasActiveMembership && (
-            <span className="text-red-500 ml-2">PRO feature</span>
-          )}
         </Button>
       </div>
 
       {/* Select document */}
-      <div className="absolute -bottom-6 -right-6 z-50">
+      <div className="absolute -bottom-5 -right-5 z-50">
         <Button
           variant="outline"
           asChild
           title="View"
-          className="bg-transparent h-8 w-8 p-0 rounded-full hover:bg-indigo-600"
+          className="bg-transparent h-6 w-6 p-0 rounded-full hover:text-white hover:bg-indigo-600"
           onClick={() => {
             router.push(`/dashboard/files/${id}`);
           }}
         >
-          <ArrowRightCircle className="h-4 w-4" />
+          <ArrowRight className="h-4 w-4" />
         </Button>
       </div>
     </div>
